requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.3
gunicorn==22.0.0
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
packaging==24.0
Werkzeug==3.0.2
zipp==3.18.1
